First pass at generic confirmation component
